# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "Aggregates news from different sources",
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "List articles with optional filters",
                "description": "Fetch a paginated list of articles with optional filters. Requires authentication.",
                "operationId": "listArticles",
                "parameters": [
                    {
                        "name": "filters",
                        "in": "query",
                        "description": "Filter criteria for articles (e.g., by keyword, date, categories and sources)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "filters[keyword]=tech&filters[date]=2025-01-01*filters[category_ids][0]=1&filters[source_ids][0]=1"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "The number of articles per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "articles": {
                                            "properties": {
                                                "article": {
                                                    "properties": {
                                                        "id": {
                                                            "description": "The ID of the article",
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "title": {
                                                            "description": "The title of the article",
                                                            "type": "string",
                                                            "example": "Example Title"
                                                        },
                                                        "url": {
                                                            "description": "The link to the article",
                                                            "type": "string",
                                                            "example": "https://www.example.com"
                                                        },
                                                        "description": {
                                                            "description": "The descritpion of the article",
                                                            "type": "string",
                                                            "example": "https://www.example.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "meta": {
                                                    "description": "Pagination metadata",
                                                    "type": "object"
                                                },
                                                "links": {
                                                    "description": "Pagination links",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article details",
                "description": "Fetch details of a specific article by its ID, including related source, category, and author information.",
                "operationId": "getArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the article",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "The ID of the article",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "description": "The title of the article",
                                            "type": "string",
                                            "example": "Example Title"
                                        },
                                        "url": {
                                            "description": "The link to the article",
                                            "type": "string",
                                            "example": "https://www.example.com"
                                        },
                                        "description": {
                                            "description": "The descritpion of the article",
                                            "type": "string",
                                            "example": "https://www.example.com"
                                        },
                                        "source": {
                                            "description": "Details of the article's source",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Tech Blog"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "category": {
                                            "description": "Details of the article's category",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Technology"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "author": {
                                            "description": "Details of the article's author",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}